Make SQLite testing class universally accessible
