+++ velodrome/fetcher/issue-events_test.go
@@ -22,12 +22,13 @@ import (
 	"time"
 
 	"k8s.io/test-infra/velodrome/sql"
+	sqltest "k8s.io/test-infra/velodrome/sql/testing"
 
 	"github.com/google/go-github/github"
 )
 
 func TestFindLatestUpdate(t *testing.T) {
-	config := SQLiteConfig{":memory:"}
+	config := sqltest.SQLiteConfig{":memory:"}
 	tests := []struct {
 		events         []sql.IssueEvent
 		expectedLatest int
@@ -75,7 +76,7 @@ func TestFindLatestUpdate(t *testing.T) {
 }
 
 func TestUpdateEvents(t *testing.T) {
-	config := SQLiteConfig{":memory:"}
+	config := sqltest.SQLiteConfig{":memory:"}
 
 	tests := []struct {
 		before []sql.IssueEvent
